Rewrite build script in C
We did quite a lot of work here, but I quite like what we ended up with.
Instead of using a single Bash script with multiple embedded Awk scripts
we now rewrote everything as a single C source with a "main"
entrypoint". The essential logic stays roughly the same with a couple of
important differences:

* Pages to be build now live in a dedicated "pages" directory (for now
  this is just a single blog post)
* All styles now live in a separate style sheet
* For now we removed the integrated Atom feed (I intend to add this at a
  later point in time)