Reverse order when iterating commits
This fixes a bug when retrieving and later assigning the modification
and creation dates of encountered files.

Due to processing commits starting from newest to oldest we would use
the actual modification date  to set the creation date first and then
proceed to erroneously use the actual creation date as the modification
date and update it successively.